requirements.txt
fastapi
sentence-transformers
openai
python-dotenv
uvicorn